Refactored code as reviews.

+++ week02/GroupAnagrams.java
@@ -5,10 +5,10 @@
 class Solution {
 
     public List<List<String>> groupAnagrams(String[] strs) {
-        Map<Integer, List<String>>anagrams = new HashMap<>();
+        Map<Long, List<String>>anagrams = new HashMap<>();
 
         for (String str : strs) {
-            int hashVal = getHashVal(str);
+            long hashVal = getHashVal(str);
             List<String>list = anagrams.getOrDefault(hashVal, new ArrayList<String>());
             list.add(str);
             anagrams.put(hashVal, list);
@@ -17,7 +17,7 @@ class Solution {
         return new ArrayList<>(anagrams.values());
     }
 
-    private int getHashVal(String s) {
+    private long getHashVal(String s) {
         long hash = 1, offset = 997, mod = 1_000_000_007;
 
         for (int i = 0; i < s.length(); i++) {
@@ -25,6 +25,6 @@ class Solution {
             hash = (hash * (ch + offset)) % mod;
         }
 
-        return (int) hash;
+        return hash;
     }
 }
